refactor(thread): share compact token formatting
- Extract a shared token formatter for goal, context, compaction, and sub-agent UI
- Update token labels to use consistent K/M suffixes across thread surfaces
- Refresh tests for the new formatting
